testingKeycloak/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
idna==2.7
pycparser==2.19
PyJWT==1.7.1
requests==2.20.1
six==1.11.0
typing==3.5.3.0
urllib3==1.24.1
